Refactor homing projectile logic into ProjectileMovementSystem
