+++ src/config/WeaponConfig.ts
@@ -46,11 +46,15 @@ export const bossWeaponPreset: WeaponConfigPreset = {
     fireRate: 1.2,
     projectileSpeed: 35.0,
     range: 18.0, // Long range
-    projectileType: 'bullet',
+    projectileType: 'homing',
     isAutoTargeting: true,
     detectionRange: 20.0, // Very long detection range
     requiresLineOfSight: false,
     leadTargetDistance: 1.0, // Aim 1.0 units in front of player ship for better boss accuracy
+    homingStrength: 0.7, // 70% homing strength - strong but dodgeable
+    homingRange: 15.0, // Can track targets within 15 units
+    homingUpdateInterval: 0.1, // Update targeting 10 times per second
+    homingTurnRate: Math.PI * 1.5, // Can turn up to 270 degrees per second
     shootingPoints: [
         { x: -0.4, y: 2 },
         { x: 0.4, y: 2 },
