+++ src/systems/index.ts
@@ -15,7 +15,7 @@ export * from './EnemyHealthUISystem'
 export * from './EnemySpawningSystem'
 export * from './GameStateSystem'
 export * from './GoalUISystem'
-export * from './HomingProjectileSystem'
+
 export * from './InputSystem'
 export * from './LevelingSystem'
 export * from './MovementSystem'
